LayoutTests/http/tests/security/credentials-iframes-expected.txt

ALERT: parent host: 127.0.0.1 iframe host: 127.0.0.1 credentials:User: same-domain-user, password: same-domain-password.
CONSOLE MESSAGE: Blocked http://127.0.0.1:8000/security/resources/cors-basic-auth.php from asking for credentials because it is a cross-origin request.
ALERT: parent host: localhost iframe host: 127.0.0.1 credentials:Authentication canceled
